requirements.bak.txt
abcd==0.3.0
addict==2.4.0
astroid==3.0.2
bcrypt==4.1.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
dill==0.3.7
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-LoginManager==1.1.6
Flask-WTF==1.2.1
idna==3.6
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
mysql-connector==2.2.9
mysql-connector-python==8.0.33
packaging==23.2
platformdirs==4.1.0
pluggy==1.3.0
protobuf==3.20.3
pylint==3.0.2
pytest==7.4.3
requests==2.31.0
six==1.16.0
tomli==2.0.1
tomlkit==0.12.3
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.16.0
WTForms==3.1.2
